backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
joblib==1.3.2
scikit-learn==1.3.0
pymongo==4.5.0